splitted databasehandler in get and mutate
